Cleaning up ringmaster for release v0.1.0
