+++ ringmaster/Cargo.toml
@@ -1,17 +1,15 @@
 [package]
 name = "ringmaster"
-version = "0.0.0"
-authors = ["Jace K. Nelson <kenneden@tamu.edu>"]
+version = "0.1.0-alpha1"
+authors = ["Mautamu <mautamu@>"]
 edition = "2018"
 
 [dependencies]
-postgres = "0.17.2"
-dotenv = "0.15.0"
+chrono = {version = "0.4.13", features = ["serde"]}
 diesel = { version = "1.4.4", features = ["postgres", "r2d2", "uuid", "chrono", "serde_json"] }
 diesel_migrations = "1.4"
-websocket = "0.24.0"
+dotenv = "0.15.0"
+postgres = "0.17.2"
+rand = "0.7.3"
 serde = "1.0"
-serde_json = "1.0"
 serde_derive = "1.0"
-rand = "*"
-chrono = {version = "0.4.13", features = ["serde"]}
